Some naming convention enhancemnets and method refactors
