+++ src/main/java/controller/Controller.java
@@ -1,7 +1,7 @@
-package sample;
+package controller;
 
 import Database.PhoneBookUsers;
-import Database.PhoneBookUsersDOA;
+import Database.PhoneBookUsersDAO;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
@@ -46,10 +46,9 @@ public class Controller {
     public void onSaveClick(){
         //check if the inputs are not null
         if (checkTheInputs(txtName,txtName,txtNumber,lblFailed)){
-            Long number = null;
-            number = Long.valueOf(txtNumber.getText().trim());
+            Long number = Long.valueOf(txtNumber.getText().trim());
             PhoneBookUsers user = new PhoneBookUsers(txtName.getText().trim(),txtSurname.getText().trim(),number);
-            PhoneBookUsersDOA createUser = new PhoneBookUsersDOA();
+            PhoneBookUsersDAO createUser = new PhoneBookUsersDAO();
 
             createUser.addUser(user);
             populateTheTable(createUser);
@@ -60,7 +59,6 @@ public class Controller {
     }
 
     public Boolean checkTheInputs(TextField name,TextField surname,TextField number,Label lbl){
-//        lblFailed.setTextFill(Paint.valueOf("#ec4d37"));
         boolean check = false;
         //CHECK IF THE INPUTS ARE NULL OR THE NUMBER IS NOT A NUMBER
         if (!name.getText().trim().isEmpty() && !surname.getText().trim().isEmpty() && !number.getText().trim().isEmpty()){
@@ -69,7 +67,6 @@ public class Controller {
                 Long.parseLong(number.getText().trim());
                 check= true;
             } catch (NumberFormatException nfe) {
-//                System.out.println("NumberFormatException: " + nfe.getMessage());
                 lbl.setText("Write a number!");
             }
         }else {
@@ -78,8 +75,7 @@ public class Controller {
         return check;
     }
 
-    private void populateTheTable(PhoneBookUsersDOA createUser){
-
+    private void populateTheTable(PhoneBookUsersDAO createUser){
         //create the table columns
         tableColumnName.setCellValueFactory(new PropertyValueFactory<PhoneBookUsers, Date>("Name"));
         tableColumnSurname.setCellValueFactory(new PropertyValueFactory<PhoneBookUsers,String>("Surname"));
@@ -94,24 +90,30 @@ public class Controller {
 
 
     public PhoneBookUsers tableClicked(MouseEvent mouseEvent) {
-
         PhoneBookUsers user = (PhoneBookUsers) userListTable.getSelectionModel().getSelectedItem();
-        lblStatus.setText("");//clear the status label
-        txtNameOld.setText(user.getName());
-        txtSurnameOld.setText(user.getSurname());
-        txtNumberOld.setText(String.valueOf(user.getNumber()));
+
+        resetLabel();
+        setTexts(user);
         return user;
     }
 
+    private void setTexts(PhoneBookUsers user) {
+        txtNameOld.setText(user.getName());
+        txtSurnameOld.setText(user.getSurname());
+        txtNumberOld.setText(String.valueOf(user.getNumber()));
+    }
+    private void resetLabel() {
+        lblStatus.setText("");//clear the status label
+    }
+
     public void onUpdateClick(MouseEvent mouseEvent) {
+        PhoneBookUsers user = (PhoneBookUsers) userListTable.getSelectionModel().getSelectedItem();
 
         if (checkTheInputs(txtNameOld,txtSurnameOld,txtNumberOld,lblStatus)){
-            PhoneBookUsers user = (PhoneBookUsers) userListTable.getSelectionModel().getSelectedItem();
             user.setName(txtNameOld.getText());
             user.setSurname(txtSurnameOld.getText());
             user.setNumber(Long.parseLong(txtNumberOld.getText()));
-
-            PhoneBookUsersDOA pu =new PhoneBookUsersDOA();
+            PhoneBookUsersDAO pu =new PhoneBookUsersDAO();
             pu.updateUser(user);
             lblStatus.setText("User updated.");
 
@@ -130,7 +132,7 @@ public class Controller {
         if (checkTheInputs(txtNameOld,txtSurnameOld,txtNumberOld,lblStatus)){
 
             PhoneBookUsers user = (PhoneBookUsers) userListTable.getSelectionModel().getSelectedItem();
-            PhoneBookUsersDOA pu =new PhoneBookUsersDOA();
+            PhoneBookUsersDAO pu =new PhoneBookUsersDAO();
             pu.deleteUser(user);
             lblStatus.setText("User Deleted.");
